Improve the documentation and allow to generate a changelog (#8)
Include improvements to the documentation and allow to generate a changelog. The changelog is generated using GitHub changelog generator through the main repository script. This pull request modifies the CI to deploy the package whenever the master branch is modified.

This pull request correctly reimplements the modifications explained in the commit of #6